inc Folder
Footer and Header removed to support CSS structure. If they must be in here please send me why, if for asthetics I'm currently ignoring it in favor of funcionality. (I can't get root folder to work how I want so it's creating errors if header isn't in the root location.)